refactor: reorganize imports and enhance OTP service classes for improved clarity and functionality
