+++ src/main/java/it/smartcommunitylab/aac/otp/provider/OtpFilterProvider.java
@@ -1,14 +1,5 @@
 package it.smartcommunitylab.aac.otp.provider;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-import javax.servlet.Filter;
-
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.util.Assert;
-
 import it.smartcommunitylab.aac.SystemKeys;
 import it.smartcommunitylab.aac.accounts.persistence.UserAccountService;
 import it.smartcommunitylab.aac.core.auth.RequestAwareAuthenticationSuccessHandler;
@@ -16,6 +7,12 @@ import it.smartcommunitylab.aac.core.provider.FilterProvider;
 import it.smartcommunitylab.aac.core.provider.ProviderConfigRepository;
 import it.smartcommunitylab.aac.internal.model.InternalUserAccount;
 import it.smartcommunitylab.aac.otp.auth.UsernameOtpAuthenticationFilter;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import javax.servlet.Filter;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.util.Assert;
 
 public class OtpFilterProvider implements FilterProvider {
